+++ packages/carp_connectivity_package/lib/connectivity_package.dart
@@ -11,7 +11,7 @@ class ConnectivitySamplingPackage extends SmartphoneSamplingPackage {
   /// Measure type for collection of nearby Bluetooth devices on a regular basis.
   ///  * Event-based (Periodic) measure - default every 10 minutes for 10 seconds.
   ///  * Uses the [Smartphone] master device for data collection.
-  ///  * Use a [PeriodicSamplingConfiguration] for configuration.
+  ///  * Use a [BluetoothScanPeriodicSamplingConfiguration] for configuration.
   static const String BLUETOOTH = "${NameSpace.CARP}.bluetooth";
 
   /// Measure type for collection of wifi information (SSID, BSSID, IP).
@@ -22,13 +22,16 @@ class ConnectivitySamplingPackage extends SmartphoneSamplingPackage {
 
   /// Measure type for Beacon ranging to detect and estimate proximity to
   /// Bluetooth beacons (e.g., iBeacon, Eddystone).
-  /// * Typically returns beacon identifiers (UUID, major, minor) and
-  ///    estimated distance or RSSI.
-  ///  * Use a [PeriodicSamplingConfiguration] for configuration.
+  /// Typically collects beacon identifiers (UUID, major, minor) and
+  /// estimated distance or RSSI.
+  ///  * Event-based (Interval) measure - default every 10 minutes.
+  ///  * Uses the [Smartphone] master device for data collection.
+  ///  * Use a [BeaconRangingPeriodicSamplingConfiguration] for configuration.
   static const String BEACON = "${NameSpace.CARP}.beacon";
 
   @override
-  DataTypeSamplingSchemeMap get samplingSchemes => DataTypeSamplingSchemeMap.from([
+  DataTypeSamplingSchemeMap get samplingSchemes =>
+      DataTypeSamplingSchemeMap.from([
         DataTypeSamplingScheme(
           CamsDataTypeMetaData(
             type: CONNECTIVITY,
@@ -59,14 +62,11 @@ class ConnectivitySamplingPackage extends SmartphoneSamplingPackage {
         DataTypeSamplingScheme(
           CamsDataTypeMetaData(
             type: BEACON,
-            displayName: "Ranging beacons in proximity",
+            displayName: "Ranging iBeacons",
             timeType: DataTimeType.POINT,
             permissions: [Permission.bluetoothScan, Permission.locationAlways],
           ),
-          PeriodicSamplingConfiguration(
-            interval: const Duration(minutes: 10),
-            duration: const Duration(seconds: 10),
-          ),
+          BeaconRangingPeriodicSamplingConfiguration(),
         ),
       ]);
 
@@ -93,15 +93,22 @@ class ConnectivitySamplingPackage extends SmartphoneSamplingPackage {
       Connectivity(),
       Bluetooth(),
       Wifi(),
+      BeaconData(region: ''),
       BluetoothScanPeriodicSamplingConfiguration(
         interval: const Duration(minutes: 10),
         duration: const Duration(seconds: 10),
       ),
+      BeaconRangingPeriodicSamplingConfiguration(
+          beaconDistance: 2, beaconRegions: [])
     ]);
 
     // registering default privacy functions
-    DataTransformerSchemaRegistry().lookup(PrivacySchema.DEFAULT)!.add(BLUETOOTH, bluetoothNameAnonymizer);
-    DataTransformerSchemaRegistry().lookup(PrivacySchema.DEFAULT)!.add(WIFI, wifiNameAnonymizer);
+    DataTransformerSchemaRegistry()
+        .lookup(PrivacySchema.DEFAULT)!
+        .add(BLUETOOTH, bluetoothNameAnonymizer);
+    DataTransformerSchemaRegistry()
+        .lookup(PrivacySchema.DEFAULT)!
+        .add(WIFI, wifiNameAnonymizer);
   }
 }
 
@@ -116,13 +123,13 @@ class ConnectivitySamplingPackage extends SmartphoneSamplingPackage {
 /// Filtering on remoteIds allows Android to scan for devices in the background
 /// without needing to be in the foreground. This is not possible on iOS.
 @JsonSerializable(includeIfNull: false, explicitToJson: true)
-class BluetoothScanPeriodicSamplingConfiguration extends PeriodicSamplingConfiguration {
+class BluetoothScanPeriodicSamplingConfiguration
+    extends PeriodicSamplingConfiguration {
   /// List of Bluetooth service UUIDs to filter the scan results.
   List<String> withServices;
 
   /// List of remote device IDs to filter the scan results.
   List<String> withRemoteIds;
-
 
   BluetoothScanPeriodicSamplingConfiguration({
     required super.interval,
@@ -132,33 +139,89 @@ class BluetoothScanPeriodicSamplingConfiguration extends PeriodicSamplingConfigu
   });
 
   @override
-  Map<String, dynamic> toJson() => _$BluetoothScanPeriodicSamplingConfigurationToJson(this);
+  Map<String, dynamic> toJson() =>
+      _$BluetoothScanPeriodicSamplingConfigurationToJson(this);
   @override
-  Function get fromJsonFunction => _$BluetoothScanPeriodicSamplingConfigurationFromJson;
-  factory BluetoothScanPeriodicSamplingConfiguration.fromJson(Map<String, dynamic> json) =>
-      FromJsonFactory().fromJson<BluetoothScanPeriodicSamplingConfiguration>(json);
+  Function get fromJsonFunction =>
+      _$BluetoothScanPeriodicSamplingConfigurationFromJson;
+  factory BluetoothScanPeriodicSamplingConfiguration.fromJson(
+          Map<String, dynamic> json) =>
+      FromJsonFactory()
+          .fromJson<BluetoothScanPeriodicSamplingConfiguration>(json);
 }
 
+/// A sampling configuration specifying how to scan for iBeacon devices.
+///
+/// The regions of the beacons to monitor are specified in the [beaconRegions]
+/// list. The [beaconDistance] is used to determine the proximity to the beacon,
+/// with a default value of 2 meters.
 @JsonSerializable(includeIfNull: false, explicitToJson: true)
-class BeaconRangingPeriodicSamplingConfiguration extends PeriodicSamplingConfiguration {
-  /// List of beacon regions to monitor and/or range using the `flutter_beacon` package.
-  List<BeaconRegion?> beaconRegions;
+class BeaconRangingPeriodicSamplingConfiguration extends SamplingConfiguration {
+  /// List of beacon regions to monitor and range.
+  List<BeaconRegion> beaconRegions;
 
-  /// When a device is within this distance from the beacon, a predefined event is triggered.
-  /// Defaults to 2 meters.
+  /// The distance in meters to consider a beacon as "in range".
   int beaconDistance;
 
   BeaconRangingPeriodicSamplingConfiguration({
-    required super.interval,
-    required super.duration,
     this.beaconRegions = const [],
     this.beaconDistance = 2,
-  });
+  }) : super();
 
   @override
-  Map<String, dynamic> toJson() => _$BeaconRangingPeriodicSamplingConfigurationToJson(this);
+  Map<String, dynamic> toJson() =>
+      _$BeaconRangingPeriodicSamplingConfigurationToJson(this);
   @override
-  Function get fromJsonFunction => _$BeaconRangingPeriodicSamplingConfigurationFromJson;
-  factory BeaconRangingPeriodicSamplingConfiguration.fromJson(Map<String, dynamic> json) =>
-      FromJsonFactory().fromJson<BeaconRangingPeriodicSamplingConfiguration>(json);
+  Function get fromJsonFunction =>
+      _$BeaconRangingPeriodicSamplingConfigurationFromJson;
+  factory BeaconRangingPeriodicSamplingConfiguration.fromJson(
+          Map<String, dynamic> json) =>
+      FromJsonFactory()
+          .fromJson<BeaconRangingPeriodicSamplingConfiguration>(json);
+}
+
+/// Beacon region to use when scanning for beacons.
+@JsonSerializable(includeIfNull: false, explicitToJson: true)
+class BeaconRegion {
+  /// A unique identifier for the beacon region.
+  /// Used to distinguish between different regions being monitored.
+  String identifier;
+
+  /// The proximity UUID of the beacon.
+  /// This is a 128-bit value used to identify a group of related beacons.
+  String uuid;
+
+  /// The major value of the beacon region (optional).
+  /// Used to further distinguish a subset of beacons within the same UUID.
+  int? major;
+
+  /// The minor value of the beacon region (optional).
+  /// Provides a finer granularity within a group of beacons identified by
+  /// the same UUID and major value.
+  int? minor;
+
+  BeaconRegion({
+    required this.identifier,
+    required this.uuid,
+    this.major,
+    this.minor,
+  });
+
+  Region toRegion() {
+    return Region(
+      identifier: identifier,
+      proximityUUID: uuid,
+      major: major,
+      minor: minor,
+    );
+  }
+
+  factory BeaconRegion.fromJson(Map<String, dynamic> json) =>
+      _$BeaconRegionFromJson(json);
+
+  Map<String, dynamic> toJson() => _$BeaconRegionToJson(this);
+
+  @override
+  String toString() =>
+      '${super.toString()}, Identifier: $identifier, UUID: $uuid, Major: $major, Minor: $minor';
 }